requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.7
async-timeout==4.0.2
attrs==21.4.0
banal==1.0.6
charset-normalizer==2.0.12
dataset==1.5.2
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
Mako==1.2.0
MarkupSafe==2.1.1
multidict==6.0.2
py-cord==2.0.0b5
python-dotenv==0.20.0
SQLAlchemy==1.4.35
yarl==1.7.2
